feat : make eye icon appear with a sound (ding) when the supervisor say "ill keep an eye on you"
feat : make the button animation play if no choice is taken for too long (the animation taht cues the player to press the right buttons)
feat : reset dialog speed after each dialog : when you press space it is set to 4 rn. Needs to be reseted so it goes back to 1 for the next dialog.
feat :